Add files via upload
Overview
This Hangman game is a simple console-based application implemented in C++. Users are prompted to guess a randomly selected country name, and they have a limited number of attempts to complete the word by guessing individual letters.

How to Play
Upon running the program, the user is welcomed to the Hangman game and instructed to guess a country name.
Each letter in the country name is represented by a star.
The user is allowed to input one letter per try, with a total of 5 attempts to guess the word.
The program provides feedback on whether the guessed letter is correct, and if so, reveals the positions of the correctly guessed letters in the word.
The game continues until the user successfully guesses the entire word or exhausts the allowed number of attempts.
If the user correctly guesses the word, they win the game. Otherwise, they lose, and the correct word is revealed.


Instructions
Clone the repository to your local machine.
Compile the C++ program using a C++ compiler.
Run the executable to start the Hangman game.
Follow on-screen instructions to guess the country name.